infra/requirements.txt
aws-cdk-lib==2.146.0
constructs>=10.0.0,<11.0.0
boto3>=1.34.0,<2.0.0
pytest>=8.2.0,<9.0.0
